Change double to decimal

+++ Core.Models/Product.cs
@@ -7,8 +7,8 @@ namespace Core.Models
     {
         public string Id { get; set; }
         public string Name { get; set; }
-        public double Quantity { get; set; }
-        public double Price { get; set; }
+        public int Quantity { get; set; }
+        public decimal Price { get; set; }
         public List<Discount> Discounts { get; set; } = new List<Discount>();
     }
 }
